# src/main/resources/api/swagger_person.yaml
swagger: '2.0'
info:
  description: 'person api'
  version: 1.0.0
  title: Swagger
host: local
basePath: /v1
tags:
  - name: person
    description: Operations about user
    externalDocs:
      description: Find out more about our person
      url: http://swagger.io
schemes:
  - https
  - http
paths:
  /person:
    get:
      tags:
        - person
      summary: Get all people in database
      description: ''
      operationId: getAllPeople
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/personDtoList'
        '400':
          description: "Invalid username supplied"
        '404':
          description: "User not found"

    post:
      tags:
        - person
      summary: Created one person in database
      description: ''
      operationId: addOnePerson
      produces:
        - application/json
      responses:
        '201':
          description: created
          schema:
            $ref: '#/definitions/ResponseDto'
        '400':
          description: "Invalid username supplied"
        '404':
          description: "User not found"

  /person/{personId}:
    get:
      tags:
        - person
      summary: Find person by ID
      description: Returns a single person
      operationId: getPersonById
      produces:
        - application/json
      parameters:
        - name: personId
          in: path
          description: ID of person to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/PersonDto'
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Person not found"
    delete:
      tags:
        - person
      summary: Deletes a person
      description: ''
      operationId: deleteOnePerson
      produces:
        - application/json
      parameters:
        - name: api_key
          in: header
          required: false
          type: string
        - name: personId
          in: path
          description: Person id to delete
          required: true
          type: integer
          format: int64
      responses:
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Person not found"

  /person/userName/{personFirstName}:
    get:
      tags:
        - person
      summary: Find person by userName
      description: Returns a list
      operationId: getPeopleByName
      produces:
        - application/json
      parameters:
        - name: personUserName
          in: path
          description: Name of people to return
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/PersonDto'
        '400':
          description: "Invalid name supplied"
        '404':
          description: "Person not found"



definitions:
  PersonDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      userName:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int32
      attributes:
        type: integer
        items:
          $ref: '#/definitions/PersonAttributes'
  PersonAttributes:
    type: object
    properties:
      idAttributes:
        type: integer
        format: int64
      age:
        type: string
      weight:
        type: string
      height:
        type: string
      nationality:
        type: string
  ResponseDto:
    type: object
    properties:
      code:
        type:  integer
        format: int
      message:
        type: string
  personDtoList:
    type: object
    properties:
      Person:
        type: array
        items:
          $ref: '#/definitions/PersonDto'


externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io